rust_decimal
Module prelude

A convenience module appropriate for glob imports (use rust_decimal::prelude::*;).

Re-exports

pub use crate::Decimal;

pub use crate::RoundingStrategy;

Traits

FromPrimitive

A generic trait for converting a number to a value.

FromStr

Parse a value from a string

MathematicalOps

Trait exposing various mathematical operations that can be applied using a Decimal. This is only present when the maths feature has been enabled.

One

Defines a multiplicative identity element for Self.

Signed

Useful functions for signed numbers (i.e. numbers that can be negative).

ToPrimitive

A generic trait for converting a value to a number.

Zero

Defines an additive identity element for Self.