Module rust_decimal::prelude

A convenience module appropriate for glob imports (use rust_decimal::prelude::*;).

Re-exports

pub use crate::Decimal;

pub use crate::RoundingStrategy;

Traits

FromPrimitive

A generic trait for converting a number to a value.

FromStr

Parse a value from a string

One

Defines a multiplicative identity element for Self.

Signed

Useful functions for signed numbers (i.e. numbers that can be negative).

ToPrimitive

A generic trait for converting a value to a number.

Zero

Defines an additive identity element for Self.