Module rust_decimal::prelude

A convenience module appropriate for glob imports (use rust_decimal::prelude::*;).

Structs

Decimal

Decimal represents a 128 bit representation of a fixed-precision decimal number. The finite set of values of type Decimal are of the form m / 10e, where m is an integer such that -296 < m < 296, and e is an integer between 0 and 28 inclusive.

Enums

RoundingStrategy

RoundingStrategy represents the different rounding strategies that can be used by round_dp_with_strategy.

Traits

FromPrimitive

A generic trait for converting a number to a value.

FromStr

Parse a value from a string

One

Defines a multiplicative identity element for Self.

Signed

Useful functions for signed numbers (i.e. numbers that can be negative).

ToPrimitive

A generic trait for converting a value to a number.

Zero

Defines an additive identity element for Self.